+++ test/me/lomin/sayang/api_test.clj
@@ -1,67 +1,4 @@
-# sayang
-
-*sayang* complects the definition of a function with its specification.
-
-## Rationale
-
-A useful summary of any function is the combination of its name, its expectations about the input and guarantees of its output.
-[clojure.spec](https://clojure.org/about/spec) provides this kind of function summary with the [s/fdef](https://clojure.org/guides/spec#_spec_ing_functions) macro.
-With *clojure.spec*, definition and specification of a function are separated concerns and there are good reasons for that decision. To quote [Alex Miller](https://groups.google.com/forum/m/#!topic/clojure/0wqvG2sef8I):
- > "There's a lot of value in separating the specs from the functions. You can put them in different places and only use just the specs (to spec an API for example) or just the functions (for production use where you don't need the specs)."
-
-Still, when you try to read or change a function, it is helpful to have the specification right next to the implementation.
-Additionally, if you need to change the number or the order of the function parameters, you only have to change it once.
-That is why you can optionally complect definition and specification of a function with *sayang*.
-
-### Syntax
-
-While *sayang* values a resemblance to the syntax of [prismatic/schema](https://github.com/plumatic/schema), it values not interfering with the functionality of [Cursive](https://cursive-ide.com/) more.
-Use the *resolve macro as*-Feature of *Cursive* to get the same tooling as function definitions with *clojure.core*.
-
-### Alternatives
-
-If you do not like the syntax or any other design decisions of *sayang*, you might want to have a look at the `defn-spec` macro of [Orchestra](https://github.com/jeaye/orchestra).
-
-## Leiningen
-
-*sayang* is available from Clojars. Add the following dependency to your *project.clj*:
-
-[![Current Version](https://clojars.org/me.lomin/sayang/latest-version.svg)](https://clojars.org/me.lomin/sayang)
-
-Generation of specifications is turned off by default. One option to activate it is by adding jvm-opts in Leiningen:
-
-```clojure
-(defproject sayang-test "0.1.0-SNAPSHOT"
-  :dependencies [[org.clojure/clojure "1.9.0"]
-                 [orchestra  "2017.11.12-1"]
-                 [org.clojure/test.check  "0.9.0"]
-                 [org.clojure/spec.alpha "0.1.143"]
-                 [me.lomin/sayang "0.2.0"]]
-  :profiles {:dev {:jvm-opts ["-Dme.lomin.sayang.activate=true"]}})
-```
-
-Alternatively, you can switch the activation toggle manually in your code:
-
-```clojure
-(alter-var-root #'me.lomin.sayang/*activate*  (constantly true))
-```
-
-Once activated, you need a runtime dependency to org.clojure/test.check.
-
-## Features
-
-* [in place function specification at definition](#specification-at-definition)
-* [partial specification](#partial-specification)
-* [support for destructuring](#support-for-destructuring)
-* [auto generate specs for multiple arities](#specification-for-multiple-arities)
-* [reference other specs](#reference-other-specs)
-* [data DSL for homogeneous collections](#data-dsl-for-homogeneous-collections)
-* [data DSL for tuples](#data-dsl-for-tuples)
-* [global switch to toggle specification](#global-switch-to-toggle-specification)
-
-
-### Specification at definition
-```clojure
+;Specification at definition
 (ns me.lomin.sayang.api-test
   (:require [clojure.test :refer :all]
             [clojure.spec.alpha :as s]
@@ -92,8 +29,8 @@ Once activated, you need a runtime dependency to org.clojure/test.check.
 (orchestra/instrument `basic-usage)
 
 (sg/sdefn int-identity {:args (s/cat :x int?)}
-  [x]
-  x)
+          [x]
+          x)
 
 (deftest specs-from-meta-map-test
   (is (= 100 (int-identity 100)))
@@ -103,9 +40,9 @@ Once activated, you need a runtime dependency to org.clojure/test.check.
                  (int-identity "100")))))
 
 (orchestra/instrument `int-identity)
-```
-### Partial specification
-```clojure
+
+; Partial specification
+
 (sg/sdefn partial-specs {:ret string?}
           [f
            [x :- int?]]
@@ -122,12 +59,12 @@ Once activated, you need a runtime dependency to org.clojure/test.check.
                  (partial-specs identity 5)))))
 
 (orchestra/instrument `partial-specs)
-```
-### Support for destructuring
-```clojure
+
+; Support for destructuring
+
 (sg/sdefn sum-first-two-elements
-  [[[a b] :- (s/tuple int? int? int?)]]
-  (+ a b))
+          [[[a b] :- (s/tuple int? int? int?)]]
+          (+ a b))
 
 (deftest support-for-destructuring-test
   (is (= 5 (sum-first-two-elements [2 3 4])))
@@ -137,9 +74,9 @@ Once activated, you need a runtime dependency to org.clojure/test.check.
                  (sum-first-two-elements [2 3])))))
 
 (orchestra/instrument `sum-first-two-elements)
-```
-### Specification for multiple arities
-```clojure
+
+; Specification for multiple arities
+
 (sg/sdefn make-magic-string {:ret string?}
           ([[x :- int?]]
            (str x "?"))
@@ -162,8 +99,8 @@ Once activated, you need a runtime dependency to org.clojure/test.check.
 
 (sg/sdefn add-map-values {:ret    int?
                           :fn result-larger-than-min-arg-value?}
-  [[{:keys [a b c]} :- map?]]
-  (+ a b c))
+          [[{:keys [a b c]} :- map?]]
+          (+ a b c))
 
 (deftest fn-spec-for-multi-arity-test
   (is (= -1 (add-map-values {:a 1 :b 0 :c -2})))
@@ -173,9 +110,9 @@ Once activated, you need a runtime dependency to org.clojure/test.check.
                  (add-map-values {:a -1 :b -2 :c -3})))))
 
 (orchestra/instrument `add-map-values)
-```
-### Reference other specs
-```clojure
+
+; Reference other specs
+
 (s/def ::number? number?)
 (sg/sdefn number-identity [[x :- ::number?]]
           x)
@@ -200,9 +137,9 @@ Once activated, you need a runtime dependency to org.clojure/test.check.
                  (call-with-7 identity)))))
 
 (orchestra/instrument `call-with-7)
-```
-### Data DSL for homogeneous collections
-```clojure
+
+; Data DSL for homogeneous collections
+
 (sg/sdefn speced-add {:ret int?}
           [[xs :- [int?]]]
           (apply + xs))
@@ -215,9 +152,9 @@ Once activated, you need a runtime dependency to org.clojure/test.check.
                  (speced-add (cons 1.0 (range 15)))))))
 
 (orchestra/instrument `speced-add)
-```
-### Data DSL for tuples
-```clojure
+
+;Data DSL for tuples
+
 (sg/sdefn sum-of-4-tuple {:ret float?}
           [[tuple :- [int? float? int? float?]]]
           (apply + tuple))
@@ -230,19 +167,3 @@ Once activated, you need a runtime dependency to org.clojure/test.check.
                  (sum-of-4-tuple [1 2 3 4])))))
 
 (orchestra/instrument `sum-of-4-tuple)
-```
-
-## Limitations
-
-* Does not work for ClojureScript, yet.
-
-## About
-
-Sayang (사양) means specification in Korean.
-
-## License
-
-Copyright © 2017 Steven Collins
-
-Distributed under the Eclipse Public License either version 1.0 or (at
-your option) any later version.